+++ .gitignore
@@ -10,5 +10,3 @@ node_modules/
 npm-debug.log*
 Thumbs.db
 .vscode/
-
-.atom-live-server.json
